# docs/api/openapi/api/openapi.yaml
openapi: 3.1.0
info:
  title: Chronos API
  description: "Shift, leave, and sickness planning \u2014 FS Phase 1 implementation."
  version: 0.2.0
paths:
...
  /api/shifts:
    get:
      tags:
      - shifts
      summary: List Shifts
      operationId: list_shifts_api_shifts_get
      parameters:
      - name: team_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      - name: period_start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Period Start
      - name: period_end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Period End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftOut'
                title: Response List Shifts Api Shifts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - shifts
      summary: Create Shift
      operationId: create_shift_api_shifts_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/shifts/{sid}:
    patch:
      tags:
      - shifts
      summary: Update Shift
      operationId: update_shift_api_shifts__sid__patch
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - shifts
      summary: Delete Shift
      operationId: delete_shift_api_shifts__sid__delete
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/shifts/{sid}/assignments:
    post:
      tags:
      - shifts
      summary: Assign User
      operationId: assign_user_api_shifts__sid__assignments_post
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/shifts/{sid}/assignments/{user_id}:
    delete:
      tags:
      - shifts
      summary: Unassign User
      operationId: unassign_user_api_shifts__sid__assignments__user_id__delete
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sid
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/shifts/{sid}/substitutes:
    get:
      tags:
      - shifts
      summary: Substitutes
      operationId: substitutes_api_shifts__sid__substitutes_get
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sid
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 5
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPublic'
                title: Response Substitutes Api Shifts  Sid  Substitutes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/shifts/plan:
    post:
      tags:
      - shifts
      summary: Run Plan
      operationId: run_plan_api_shifts_plan_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/reports/shifts:
    get:
      tags:
      - reports
      summary: Shifts Summary
      operationId: shifts_summary_api_reports_shifts_get
      parameters:
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Start
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date
          title: End
      - name: team_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftSummaryRow'
                title: Response Shifts Summary Api Reports Shifts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AssignmentIn:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - user_id
      title: AssignmentIn
    AssignmentOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        shift_id:
          type: string
          format: uuid
          title: Shift Id
        user_id:
          type: string
          format: uuid
          title: User Id
        status:
          type: string
          title: Status
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - id
      - shift_id
      - user_id
      - status
      - note
      title: AssignmentOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PlanIn:
      properties:
        team_id:
          type: string
          format: uuid
          title: Team Id
        period_start:
          type: string
          format: date
          title: Period Start
        period_end:
          type: string
          format: date
          title: Period End
        required_per_shift:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: Required Per Shift
          default: 1
        holidays:
          items:
            type: string
            format: date
          type: array
          title: Holidays
      type: object
      required:
      - team_id
      - period_start
      - period_end
      title: PlanIn
    PlanOut:
      properties:
        shifts_created:
          type: integer
          title: Shifts Created
        assignments_created:
          type: integer
          title: Assignments Created
        unassigned:
          items:
            type: string
          type: array
          title: Unassigned
      type: object
      required:
      - shifts_created
      - assignments_created
      - unassigned
      title: PlanOut
...
    ShiftIn:
      properties:
        team_id:
          type: string
          format: uuid
          title: Team Id
        service_date:
          type: string
          format: date
          title: Service Date
        shift_type:
          $ref: '#/components/schemas/ShiftType'
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          type: string
          format: time
          title: End Time
        required_headcount:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: Required Headcount
          default: 1
        project_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
      type: object
      required:
      - team_id
      - service_date
      - shift_type
      - start_time
      - end_time
      title: ShiftIn
    ShiftOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        team_id:
          type: string
          format: uuid
          title: Team Id
        service_date:
          type: string
          format: date
          title: Service Date
        shift_type:
          type: string
          title: Shift Type
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          type: string
          format: time
          title: End Time
        required_headcount:
          type: integer
          title: Required Headcount
        project_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
        assignments:
          items:
            $ref: '#/components/schemas/AssignmentOut'
          type: array
          title: Assignments
          default: []
      type: object
      required:
      - id
      - team_id
      - service_date
      - shift_type
      - start_time
      - end_time
      - required_headcount
      - project_id
      title: ShiftOut
    ShiftSummaryRow:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        user_name:
          type: string
          title: User Name
        early_count:
          type: integer
          title: Early Count
        late_count:
          type: integer
          title: Late Count
        other_count:
          type: integer
          title: Other Count
      type: object
      required:
      - user_id
      - user_name
      - early_count
      - late_count
      - other_count
      title: ShiftSummaryRow
    ShiftType:
      type: string
      enum:
      - EARLY
      - LATE
      - OTHER
      title: ShiftType
    ShiftUpdate:
      properties:
        start_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: End Time
        required_headcount:
          anyOf:
          - type: integer
            maximum: 20.0
            minimum: 1.0
          - type: 'null'
          title: Required Headcount
        project_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
      type: object
      title: ShiftUpdate
...
    UserPublic:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        role:
          type: string
          title: Role
        locale:
          type: string
          title: Locale
        time_zone:
          type: string
          title: Time Zone
        department_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Department Id
        team_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
        must_rotate_password:
          type: boolean
          title: Must Rotate Password
          default: false
      type: object
      required:
      - id
      - email
      - role
      - locale
      - time_zone
      title: UserPublic
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...